+++ src/contracts/index.ts
@@ -1,4 +1,5 @@
 export * from './asset';
 export * from './breadth';
 export * from './environment';
+export * from './integration';
 export * from './semantic';